requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.45.0
apify-client==1.8.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.1